needed to return an exit status for build_and_commit function so invoker knows if it failed or not

+++ scripts/update_babbler.sh
@@ -136,7 +136,10 @@ do
   git commit -m "$msg"
 
   # compile application, run test harness, and run all input files
-  build_and_test
+  if ! build_and_test; then
+    failed=true
+    break
+  fi
 
   # ensure a clean repository (you can't run this command enough here)
   git clean -xdf &> /dev/null
